클래스 파일 정리 및 구분
